Rename pojo -> model for ease of understanding
