+++ src/main/java/org/amoseman/budgetingbackend/application/BudgetingApplication.java
@@ -8,8 +8,8 @@ import io.dropwizard.core.setup.Environment;
 import org.amoseman.budgetingbackend.application.auth.*;
 import org.amoseman.budgetingbackend.application.auth.hashing.ArgonHash;
 import org.amoseman.budgetingbackend.application.auth.hashing.Hash;
-import org.amoseman.budgetingbackend.application.pojo.DAOs;
-import org.amoseman.budgetingbackend.application.pojo.Services;
+import org.amoseman.budgetingbackend.application.model.DAOs;
+import org.amoseman.budgetingbackend.application.model.Services;
 import org.amoseman.budgetingbackend.dao.AccountDAO;
 import org.amoseman.budgetingbackend.dao.FinanceRecordDAO;
 import org.amoseman.budgetingbackend.dao.BucketDAO;
@@ -19,7 +19,7 @@ import org.amoseman.budgetingbackend.dao.impl.sql.BucketDAOImpl;
 import org.amoseman.budgetingbackend.database.DatabaseConnection;
 import org.amoseman.budgetingbackend.database.impl.sql.sqlite.DatabaseConnectionImpl;
 import org.amoseman.budgetingbackend.exception.mapping.*;
-import org.amoseman.budgetingbackend.pojo.account.Account;
+import org.amoseman.budgetingbackend.model.account.Account;
 import org.amoseman.budgetingbackend.resource.AccountResource;
 import org.amoseman.budgetingbackend.resource.FinanceRecordResource;
 import org.amoseman.budgetingbackend.resource.BucketResource;
