+++ src/main/java/org/amoseman/budgetingbackend/dao/impl/sql/AccountDAOImpl.java
@@ -4,7 +4,7 @@ import org.amoseman.budgetingbackend.dao.AccountDAO;
 import org.amoseman.budgetingbackend.database.DatabaseConnection;
 import org.amoseman.budgetingbackend.exception.AccountAlreadyExistsException;
 import org.amoseman.budgetingbackend.exception.AccountDoesNotExistException;
-import org.amoseman.budgetingbackend.pojo.account.Account;
+import org.amoseman.budgetingbackend.model.account.Account;
 import org.jooq.*;
 import java.util.List;
 import java.util.Optional;
