+++ src/main/java/org/amoseman/budgetingbackend/dao/impl/sql/BucketDAOImpl.java
@@ -4,7 +4,7 @@ import org.amoseman.budgetingbackend.dao.BucketDAO;
 import org.amoseman.budgetingbackend.database.DatabaseConnection;
 import org.amoseman.budgetingbackend.exception.BucketAlreadyExistsException;
 import org.amoseman.budgetingbackend.exception.BucketDoesNotExistException;
-import org.amoseman.budgetingbackend.pojo.bucket.Bucket;
+import org.amoseman.budgetingbackend.model.bucket.Bucket;
 import org.jooq.*;
 
 import java.util.List;
