+++ src/main/java/org/amoseman/budgetingbackend/model/account/op/CreateAccount.java
@@ -1,4 +1,4 @@
-package org.amoseman.budgetingbackend.pojo.account.op;
+package org.amoseman.budgetingbackend.model.account.op;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
