+++ src/main/java/org/amoseman/budgetingbackend/model/bucket/Bucket.java
@@ -1,6 +1,6 @@
-package org.amoseman.budgetingbackend.pojo.bucket;
+package org.amoseman.budgetingbackend.model.bucket;
 
-import org.amoseman.budgetingbackend.pojo.OwnedEntity;
+import org.amoseman.budgetingbackend.model.OwnedEntity;
 
 import java.beans.ConstructorProperties;
 import java.time.LocalDateTime;
