+++ src/main/java/org/amoseman/budgetingbackend/model/record/Expense.java
@@ -1,7 +1,7 @@
-package org.amoseman.budgetingbackend.pojo.record;
+package org.amoseman.budgetingbackend.model.record;
 
 import org.amoseman.budgetingbackend.exception.NegativeValueException;
-import org.amoseman.budgetingbackend.pojo.record.info.ExpenseInfo;
+import org.amoseman.budgetingbackend.model.record.info.ExpenseInfo;
 
 import java.beans.ConstructorProperties;
 import java.time.LocalDateTime;
