+++ src/main/java/org/amoseman/budgetingbackend/service/impl/BucketServiceImpl.java
@@ -5,10 +5,10 @@ import org.amoseman.budgetingbackend.dao.BucketDAO;
 import org.amoseman.budgetingbackend.exception.BucketAlreadyExistsException;
 import org.amoseman.budgetingbackend.exception.BucketDoesNotExistException;
 import org.amoseman.budgetingbackend.exception.TotalBucketShareExceededException;
-import org.amoseman.budgetingbackend.pojo.record.Expense;
-import org.amoseman.budgetingbackend.pojo.record.Income;
-import org.amoseman.budgetingbackend.pojo.bucket.Bucket;
-import org.amoseman.budgetingbackend.pojo.bucket.BucketInfo;
+import org.amoseman.budgetingbackend.model.record.Expense;
+import org.amoseman.budgetingbackend.model.record.Income;
+import org.amoseman.budgetingbackend.model.bucket.Bucket;
+import org.amoseman.budgetingbackend.model.bucket.BucketInfo;
 import org.amoseman.budgetingbackend.service.BucketService;
 import org.amoseman.budgetingbackend.util.Now;
 import org.amoseman.budgetingbackend.util.SplitCurrency;
