+++ src/main/java/org/amoseman/budgetingbackend/util/CurrencySplitter.java
@@ -1,6 +1,6 @@
 package org.amoseman.budgetingbackend.util;
 
-import org.amoseman.budgetingbackend.pojo.bucket.Bucket;
+import org.amoseman.budgetingbackend.model.bucket.Bucket;
 
 import java.util.ArrayList;
 import java.util.Arrays;
