+++ src/test/java/org/amoseman/budgetingbackend/application/BudgetingApplicationTest.java
@@ -8,10 +8,10 @@ import org.amoseman.InitTestConfiguration;
 import org.amoseman.InitTestDatabase;
 import org.amoseman.StatusTest;
 import org.amoseman.StatusTester;
-import org.amoseman.budgetingbackend.pojo.account.op.CreateAccount;
-import org.amoseman.budgetingbackend.pojo.bucket.BucketInfo;
-import org.amoseman.budgetingbackend.pojo.record.info.ExpenseInfo;
-import org.amoseman.budgetingbackend.pojo.record.info.IncomeInfo;
+import org.amoseman.budgetingbackend.model.account.op.CreateAccount;
+import org.amoseman.budgetingbackend.model.bucket.BucketInfo;
+import org.amoseman.budgetingbackend.model.record.info.ExpenseInfo;
+import org.amoseman.budgetingbackend.model.record.info.IncomeInfo;
 import org.glassfish.jersey.client.JerseyClientBuilder;
 import org.glassfish.jersey.client.authentication.HttpAuthenticationFeature;
 import org.junit.jupiter.api.AfterAll;
